+++ src/EndpointIdentifier.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Tomaj\NetteApi;
 
+use InvalidArgumentException;
+
 class EndpointIdentifier implements EndpointInterface
 {
     private $method;
@@ -14,9 +16,19 @@ class EndpointIdentifier implements EndpointInterface
 
     private $apiAction;
 
-    public function __construct(string $method, int $version, string $package, ?string $apiAction = null)
+    /**
+     * @param string $method example: "GET", "POST", "PUT", "DELETE"
+     * @param string|int $version Version must have semantic numbering. For example "1", "1.1", "0.13.2" etc.
+     * @param string $package example: "users"
+     * @param string|null $apiAction example: "query"
+     */
+    public function __construct(string $method, $version, string $package, ?string $apiAction = null)
     {
+        $version = (string) $version;
         $this->method = strtoupper($method);
+        if (strpos($version, '/') !== false) {
+            throw new InvalidArgumentException('Version must have semantic numbering. For example "1", "1.1", "0.13.2" etc.');
+        }
         $this->version = $version;
         $this->package = $package;
         $this->apiAction = $apiAction;
@@ -27,7 +39,7 @@ class EndpointIdentifier implements EndpointInterface
         return $this->method;
     }
 
-    public function getVersion(): int
+    public function getVersion(): string
     {
         return $this->version;
     }