+++ tests/Handler/ApiListingHandlerTest.php
@@ -6,6 +6,7 @@ namespace Tomaj\NetteApi\Test\Handler;
 
 use Nette\Application\LinkGenerator;
 use Nette\Application\Routers\SimpleRouter;
+use Nette\DI\Container;
 use Nette\Http\UrlScript;
 use PHPUnit\Framework\TestCase;
 use Tomaj\NetteApi\ApiDecider;
@@ -18,25 +19,33 @@ use Tomaj\NetteApi\Link\ApiLink;
 
 class ApiListingHandlerTest extends TestCase
 {
+    /** @var Container */
+    private $container;
+
+    protected function setUp(): void
+    {
+        $this->container = new Container();
+    }
+
     public function testDefaultHandle()
     {
         $linkGenerator = new LinkGenerator(new SimpleRouter([]), new UrlScript('http://test/'));
         $apiLink = new ApiLink($linkGenerator);
 
-        $apiDecider = new ApiDecider();
+        $apiDecider = new ApiDecider($this->container);
         $apiDecider->addApi(
-            new EndpointIdentifier('POST', 2, 'comments', 'list'),
+            new EndpointIdentifier('POST', '2', 'comments', 'list'),
             new AlwaysOkHandler(),
             new NoAuthorization()
         );
 
         $apiDecider->addApi(
-            new EndpointIdentifier('GET', 2, 'endpoints'),
+            new EndpointIdentifier('GET', '2', 'endpoints'),
             new ApiListingHandler($apiDecider, $apiLink),
             new NoAuthorization()
         );
 
-        $result = $apiDecider->getApi('GET', 2, 'endpoints');
+        $result = $apiDecider->getApi('GET', '2', 'endpoints');
         $handler = $result->getHandler();
 
         $response = $handler->handle([]);
@@ -50,20 +59,20 @@ class ApiListingHandlerTest extends TestCase
         $linkGenerator = new LinkGenerator(new SimpleRouter([]), new UrlScript('http://test/'));
         $apiLink = new ApiLink($linkGenerator);
 
-        $apiDecider = new ApiDecider();
+        $apiDecider = new ApiDecider($this->container);
         $apiDecider->addApi(
-            new EndpointIdentifier('POST', 1, 'comments', 'list'),
+            new EndpointIdentifier('POST', '1', 'comments', 'list'),
             new EchoHandler(),
             new NoAuthorization()
         );
 
         $apiDecider->addApi(
-            new EndpointIdentifier('GET', 1, 'endpoints'),
+            new EndpointIdentifier('GET', '1', 'endpoints'),
             new ApiListingHandler($apiDecider, $apiLink),
             new NoAuthorization()
         );
 
-        $result = $apiDecider->getApi('GET', 1, 'endpoints');
+        $result = $apiDecider->getApi('GET', '1', 'endpoints');
         $handler = $result->getHandler();
 
         $response = $handler->handle([]);