+++ tests/Handler/OpenApiHandlerTest.php
@@ -6,6 +6,7 @@ namespace Tomaj\NetteApi\Test\Handler;
 
 use Nette\Application\LinkGenerator;
 use Nette\Application\Routers\SimpleRouter;
+use Nette\DI\Container;
 use Nette\Http\UrlScript;
 use PHPUnit\Framework\TestCase;
 use Tomaj\NetteApi\ApiDecider;
@@ -17,26 +18,34 @@ use Nette\Http\Request;
 
 class OpenApiHandlerTest extends TestCase
 {
+    /** @var Container */
+    private $container;
+
+    protected function setUp(): void
+    {
+        $this->container = new Container();
+    }
+
     public function testHandlerWithMultipleResponseSchemas()
     {
         $linkGenerator = new LinkGenerator(new SimpleRouter([]), new UrlScript('http://test/'));
         $apiLink = new ApiLink($linkGenerator);
         $request = new Request(new UrlScript('http://test/'));
 
-        $apiDecider = new ApiDecider();
+        $apiDecider = new ApiDecider($this->container);
         $apiDecider->addApi(
-            new EndpointIdentifier('GET', 1, 'test'),
+            new EndpointIdentifier('GET', '1', 'test'),
             new MultipleOutputTestHandler(),
             new NoAuthorization()
         );
 
         $apiDecider->addApi(
-            new EndpointIdentifier('GET', 1, 'docs', 'open-api'),
+            new EndpointIdentifier('GET', '1', 'docs', 'open-api'),
             new OpenApiHandler($apiDecider, $apiLink, $request),
             new NoAuthorization()
         );
 
-        $result = $apiDecider->getApi('GET', 1, 'docs', 'open-api');
+        $result = $apiDecider->getApi('GET', '1', 'docs', 'open-api');
         $handler = $result->getHandler();
 
         $response = $handler->handle(['format' => 'json']);