+++ CHANGELOG.md
@@ -4,10 +4,25 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased][unreleased]
 
+## 3.0.0
+
+### Changed
+* Support for semantic versioning api.
+* [BC] DefaultHandler response code 404 instead 400
+* [BC] Added Container to API Decider
+* [BC] Output Configurator, Allows different methods for output configuration. Needs to be added to config services.
+* [BC] Error handler, Allows for custom error handling of handle method. Needs to be added to config services.
+* Query configurator rework
+
+### Added
+* CorsPreflightHandlerInterface - resolve multiple service registered handler error
+* Lazy API handlers
+
 ## 2.12.0
 
 ### Added
 * Button to copy `Body` content in api console
+* Ability to disable schema validation and provide additional error info with get parameters. 
 
 ### Changed
 * Handler tag wrapper has changed class from `btn` to `label`